Move global data operations to separate folder
