+++ PyStemmusScope/global_data/__init__.py
@@ -1,0 +1,5 @@
+"""Module for operations related to the 'global' datasets."""
+from . import global_data_selection
+
+
+__all__ = ["global_data_selection"]
