+++ tests/test_global_data_selection.py
@@ -2,7 +2,7 @@ from pathlib import Path
 import numpy as np
 import pytest
 from PyStemmusScope import forcing_io
-from PyStemmusScope import global_data_selection as gds
+from PyStemmusScope.global_data import global_data_selection as gds
 from . import data_folder
